+++ Plugins/CoreModel/src/main/java/de/fraunhofer/iosb/ilt/frostserver/plugin/coremodel/TableImpDatastreams.java
@@ -364,7 +364,7 @@ public class TableImpDatastreams extends StaTableAbstract<TableImpDatastreams> {
                                     uom.setSymbol(value);
                                 }
                                 case "definition" -> {
-                                    uom.setSymbol(value);
+                                    uom.setDefinition(value);
                                 }
                                 default -> {
                                     // Do nothing.